Reorganising files, fixed price and rating data type, modified admin dashboard, added new report, etc...

+++ eCourses/eCourses.Mobile/eCourses.Mobile/ViewModels/MyCourses/UsersBoughtCoursesDetailVM.cs
@@ -117,7 +117,7 @@ namespace eCourses.Mobile.ViewModels.MyCourses
         public async void GetReview()
         {
             await SetCourseReview(Course.CourseID);
-            Rating = courseReview != null ? courseReview.Rating : 0;
+            Rating = (decimal)(courseReview != null ? courseReview.Rating : 0);
         }
 
         private async Task SetCourseReview(int CourseID)
